Refactoring PostComponent
한뭉탱이로 놓여져있던 컴포넌트를 나눠서 정리함 더미데이터  Context로 전역으로 관리
